backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.39.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
weasyprint==62.3
qrcode[pil]==7.4.2
supabase==2.9.1
stripe>=8.0.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
